requirements.txt
fastapi[standard]

langchain
langchain_community
langchain_cohere
langchain_chroma
chromadb
cohere
pandas
bs4
pypdf

sqlalchemy
mysql-connector-python
mysqlclient
aiomysql

twilio

mercadopago

python-jose
passlib
bcrypt
tqdm